+++ src/lib/plugins.ts
@@ -8,6 +8,8 @@
  */
 import { seoPlugin } from '@aphexcms/plugin-seo';
 import { colorPickerPlugin } from '@aphexcms/plugin-color-picker';
+import { notifyPlugin } from './plugins/notify-plugin';
+import { formsPlugin } from './forms/forms-plugin';
 
 export const plugins = [
 	seoPlugin({
@@ -25,5 +27,7 @@ export const plugins = [
 			return `/blog/${slug}`;
 		}
 	}),
-	colorPickerPlugin()
+	colorPickerPlugin(),
+	notifyPlugin,
+	formsPlugin
 ];